requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
fastapi==0.115.0
h11==0.16.0
idna==3.10
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.14.1
uvicorn==0.30.6
